split up and add some more tests
