+++ tests/test_pipeline.py
@@ -2,8 +2,6 @@
 
 import tempfile
 
-import numpy as np
-
 import btk
 from btk.survey import Survey
 
@@ -112,167 +110,3 @@ def test_pipeline(data_dir):
 
         assert deblend_batch.batch_size == deblend_batch2.batch_size
         assert deblend_batch.image_size == deblend_batch2.image_size
-
-
-def test_cosmos_generator(data_dir):
-    """Test the pipeline as a whole for a single deblender."""
-    cosmos_catalog_paths = [
-        data_dir / "cosmos" / "real_galaxy_catalog_23.5_example.fits",
-        data_dir / "cosmos" / "real_galaxy_catalog_23.5_example_fits.fits",
-    ]
-    cosmos_catalog_files = [p.as_posix() for p in cosmos_catalog_paths]
-    catalog = btk.catalog.CosmosCatalog.from_file(cosmos_catalog_files)
-
-    _ = catalog.get_raw_catalog()
-
-    survey: Survey = btk.survey.get_surveys("LSST")
-    fltr = survey.get_filter("r")
-    assert hasattr(fltr, "psf")
-
-    stamp_size = 24.0
-    max_shift = 1.0
-    max_n_sources = 2
-    sampling_function = btk.sampling_functions.DefaultSampling(
-        max_number=max_n_sources,
-        min_number=1,
-        stamp_size=stamp_size,
-        max_shift=max_shift,
-        min_mag=20,
-        max_mag=21,
-        seed=SEED,
-        mag_name="MAG",
-    )
-
-    batch_size = 10
-
-    draw_generator = btk.draw_blends.CosmosGenerator(
-        catalog,
-        sampling_function,
-        survey,
-        batch_size=batch_size,
-        stamp_size=stamp_size,
-        njobs=1,
-        add_noise="all",
-        seed=SEED,
-        gal_type="real",
-    )
-
-    # generate batch 100 blend catalogs and images.
-    blend_batch = next(draw_generator)
-    assert len(blend_batch.catalog_list) == batch_size
-    assert blend_batch.blend_images.shape == (batch_size, 6, stamp_size / 0.2, stamp_size / 0.2)
-    iso_shape = (batch_size, max_n_sources, 6, stamp_size / 0.2, stamp_size / 0.2)
-    assert blend_batch.isolated_images.shape == iso_shape
-
-
-def test_sep(data_dir):
-    """Check we always detect single bright objects."""
-
-    catalog_file = data_dir / "input_catalog.fits"
-    catalog = btk.catalog.CatsimCatalog.from_file(catalog_file)
-    survey: Survey = btk.survey.get_surveys("LSST")
-
-    # single bright galaxy=
-    sampling_function = btk.sampling_functions.DefaultSampling(
-        max_number=1,
-        min_number=1,
-        stamp_size=24.0,
-        max_shift=1.0,
-        min_mag=0,
-        max_mag=21,
-        seed=SEED,
-    )
-
-    assert np.sum((catalog.table["i_ab"] > 0) & (catalog.table["i_ab"] < 21)) > 100
-
-    batch_size = 100
-
-    draw_generator = btk.draw_blends.CatsimGenerator(
-        catalog,
-        sampling_function,
-        survey,
-        batch_size=batch_size,
-        stamp_size=24.0,
-        njobs=1,
-        add_noise="all",
-        seed=SEED,
-    )
-
-    blend_batch = next(draw_generator)
-    deblender = btk.deblend.SepSingleBand(max_n_sources=1, thresh=3, use_band=2)
-    deblend_batch = deblender(blend_batch, njobs=1)
-
-    matcher = btk.match.PixelHungarianMatcher(pixel_max_sep=5.0)
-
-    true_catalog_list = blend_batch.catalog_list
-    pred_catalog_list = deblend_batch.catalog_list
-    matching = matcher(true_catalog_list, pred_catalog_list)  # matching object
-    tp, t, p = matching.tp, matching.t, matching.p
-
-    recall = btk.metrics.detection.Recall(batch_size)
-    precision = btk.metrics.detection.Precision(batch_size)
-
-    assert recall(tp, t, p) > 0.95
-    assert precision(tp, t, p) > 0.95
-
-
-def test_scarlet(data_dir):
-    """Check scarlet deblender implementation runs without too many failures."""
-
-    max_n_sources = 3
-    stamp_size = 24.0
-    seed = 0
-    max_shift = 2.0  # shift is only 2 arcsecs -> 10 pixels, so blends are likely.
-
-    catalog = btk.catalog.CatsimCatalog.from_file(data_dir / "input_catalog.fits")
-    sampling_function = btk.sampling_functions.DefaultSampling(
-        max_number=max_n_sources,
-        min_number=max_n_sources,  # always 3 sources in every blend.
-        stamp_size=stamp_size,
-        max_shift=max_shift,
-        min_mag=24,
-        max_mag=25,
-        seed=seed,
-    )
-    LSST = btk.survey.get_surveys("LSST")
-
-    batch_size = 10
-
-    draw_generator = btk.draw_blends.CatsimGenerator(
-        catalog,
-        sampling_function,
-        LSST,
-        batch_size=batch_size,
-        stamp_size=stamp_size,
-        njobs=1,
-        add_noise="all",
-        seed=seed,  # use same seed here
-    )
-
-    blend_batch = next(draw_generator)
-    deblender = btk.deblend.Scarlet(max_n_sources)
-    deblend_batch = deblender(blend_batch, reference_catalogs=blend_batch.catalog_list)
-    n_failures = np.sum([len(cat) == 0 for cat in deblend_batch.catalog_list], axis=0)
-    assert n_failures <= 3
-
-
-def test_density_sampling(data_dir):
-    """Test new density sampling function."""
-    catalog = btk.catalog.CatsimCatalog.from_file(data_dir / "input_catalog.fits")
-
-    sampling_function = btk.sampling_functions.DensitySampling(
-        max_number=50,
-        min_number=1,
-        stamp_size=24,
-        max_shift=10,
-        min_mag=-np.inf,
-        max_mag=27.3,
-        seed=SEED,
-    )
-
-    for _ in range(10):
-        blends = sampling_function(catalog.get_raw_catalog())
-
-        assert 1 <= len(blends) <= 50
-        assert np.all(blends["ra"] >= -10) & np.all(blends["ra"] <= 10)
-        assert np.all(blends["dec"] >= -10) & np.all(blends["dec"] <= 10)
